Update example METS files to include related items
This update adds a new METS sample file, 86457-mets-2.0.xml to the
sample XML outputs. This contains example mappings for, in particular
publication-publication relationships.

Updates have been made to the Symplectic -> ORA mappings file to
describe the mappings which created that relationship.

This update also adds a number of OCFL related changes: these are
very relevant to other areas of ORA Sword2 development, but do
not affect Symplectic Ltd.